# performance and type stability fixes for geometries