Improves type stability where possible by keeping enums as types in `Val{wkbLineString}` and by forcing child object types from `getgeom`.

~~Also adds a low allocation `getpoint` method.~~

Edit: review when I get tests passing, can't always test on a local machine at present