@rafaqz 2022-08-11
The failures are an unrelated issue with GDAL_jll on mac. Otherwise this is good to go.