@rafaqz 2022-09-15
The abstract types in this PR are getting out of control, because we dont get dimensionality or object type from the enum, we have to use huge `Union` types to define everything in groups we can dispatch on.

It might be nicer to use generic type parameters to indicate the kinds of objects and if they are 2D, 3D or have M fields, and delete all the `Union` groups for e.g. all 3d object types, all M object types etc.

So, something like:
```juila
abstract type AbstractGeometry{E,T,D,M} end
struct Geometry{E,T,D,M} <: AbstractGeometry{E,T,D,M}
    ....
end
struct IGeometry{E,T,D,M} <: AbstractGeometry{E,T,D,M} 
    ...
end
# With default constructors like this:
Geometry{E}(x) where E = Geometry{W,geomtype(W),dimensionality(W),hasm(W)}(x)
```

Then, for example, the types would be:
```julia
Geometry{wkbLineStringZM,LineString,3,HasM}
Geometry{wkbPoint,Point,2,NoM}
```

And we can just use dispatch to match any of the type paramters instead of using so many huge `Union` types.

Any comments on this @yeesian @evetion ?