@rafaqz 2022-09-16
Probably that type shouldn't exist and we could make sure of that in the inner constructor. 

We could also not include the enum in the type at all, and generate it when needed.