@yeesian 2022-10-04
> We could also not include the enum in the type at all, and generate it when needed.

I think I like this more. How about something like

```julia
Geometry{wkbLineStringZM,3}
Geometry{wkbPoint,2}
```

And we can generate (i) `LineString` and `HasM` from `(wkbLineStringZM, 3)` and (ii) `Point` and `NoM` from `(wkbPoint,2)`.