@rafaqz 2022-10-04
I actually meant the other way round:

```julia
Geometry{LineString,3,HasM}
```

As `wkbLineString` has no use in dispatch, so we would still need all the big unions I have defined.

We can get `wkbLineString` Frome the other three as we don't need it for dispatch.

It might be better to merge this as-is when I fix the tests, and think about this separately. I'm not sure.