@yeesian 2022-10-05
> I actually meant the other way round

Ah yes sorry you're right!

> It might be better to merge this as-is when I fix the tests, and think about this separately. I'm not sure.

Yeah I think it'll be a big change regardless of the outcome of the proposal, and agree about thinking about it separately. Happy with this PR once the tests are fixed