@rafaqz 2022-10-08
I think this is good to go. 

I found a few existing type-related problems while doing this: `flattento2d!` and `setcoorddim!` etc don't actually work because we don't update the type. Even though its and in place function we will have to use the return value in future. I'll make some issues and fix elsewhere.