@rafaqz 2022-10-22
Ok, made a few changes. Measures handling was really not ready so I wound that back a little. This means `GI.m` and `ismeasured` are tested for the `true` case as there isn't an obvious way to make measured objects. 

I've tested most of the types more extensively for `is3d`, for type parameters and child type parameters. The missing lines now (besides `GI.m` and `ismeasured`) seem to be errors in coverage, like constants counted as lines of code that don't run.