@yeesian 2022-10-23
I'm going to consider the review by @evetion in https://github.com/yeesian/ArchGDAL.jl/pull/316#pullrequestreview-1072118118 as resolved by https://github.com/yeesian/ArchGDAL.jl/pull/316#issuecomment-1245931594 and the subsequent commits to increase test coverage.